server/requirements.txt
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.20.0
pymongo==4.5.0
pydantic-settings==2.2.1
python-dotenv==0.21.1
SQLAlchemy==2.0.13
loguru==0.7.2
python-docx==1.1.0
requests==2.31.0
passlib==1.7.4
python-jose[cryptography]